+++ modules/apps/29-fee/keeper/escrow.go
@@ -229,6 +229,9 @@ func (k Keeper) RefundFeesOnChannelClosure(ctx sdk.Context, portID, channelID st
 		k.DeleteFeesInEscrow(cacheCtx, identifiedPacketFee.PacketId)
 	}
 
+	// NOTE: The context returned by CacheContext() refers to a new EventManager, so it needs to explicitly set events to the original context.
+	ctx.EventManager().EmitEvents(cacheCtx.EventManager().Events())
+
 	// write the cache
 	writeFn()